Separate create-foxglove-extension package from scripts package
