+++ packages/create-foxglove-extension/README.md
@@ -1,0 +1,7 @@
+# create-foxglove-extension
+
+[![npm version](https://img.shields.io/npm/v/create-foxglove-extension.svg?style=flat)](https://www.npmjs.com/package/create-foxglove-extension)
+
+```sh
+npm init foxglove-extension@latest helloworld
+```
