+++ packages/create-foxglove-extension/src/create.test.ts
