+++ packages/create-foxglove-extension/src/log.ts
