+++ packages/create-foxglove-extension/template/.eslintrc.yaml
