+++ packages/create-foxglove-extension/template/.gitignore
