+++ packages/create-foxglove-extension/template/.prettierrc.yaml
