+++ packages/create-foxglove-extension/template/CHANGELOG.md
