+++ packages/create-foxglove-extension/template/LICENSE
