+++ packages/create-foxglove-extension/template/README.md
