+++ packages/create-foxglove-extension/template/package.json
