+++ packages/create-foxglove-extension/template/src/ExamplePanel.tsx
