+++ packages/create-foxglove-extension/template/src/index.ts
