+++ packages/foxglove-extension/src/bin/foxglove-extension.ts
