+++ packages/foxglove-extension/src/build.ts
