+++ packages/foxglove-extension/src/extensions.ts
