+++ packages/foxglove-extension/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./build";
-export * from "./create";
 export * from "./package";
