+++ packages/foxglove-extension/src/log.ts
@@ -1,0 +1,8 @@
+export function info(message: unknown, ...args: unknown[]): void {
+  console.log(message, ...args);
+}
+
+export function fatal(message: unknown, ...args: unknown[]): void {
+  console.error(message, ...args);
+  process.exit(1);
+}
