+++ packages/foxglove-extension/src/package.ts
