+++ packages/foxglove-extension/src/webpackConfigExtension.ts
