+++ packages/foxglove-extension/tsconfig/tsconfig.json
