--- tsconfig.json
@@ -1,32 +1,0 @@
-{
-  "include": ["./src/**/*.ts"],
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es2020",
-    "lib": ["es2020"],
-    "rootDir": "./src",
-    "outDir": "./dist",
-
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "isolatedModules": true,
-    "allowJs": false,
-    "newLine": "lf",
-    "resolveJsonModule": true,
-    "skipLibCheck": true,
-
-    "declaration": true,
-    "declarationMap": true,
-    "sourceMap": true,
-
-    // be as strict as possible, but no stricter
-    "strict": true,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitAny": true,
-    "noImplicitReturns": true,
-    "noUncheckedIndexedAccess": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "forceConsistentCasingInFileNames": true
-  }
-}
